requirements.txt
behave
browserstack-local
paver
